+++ CHANGELOG.md
@@ -2,12 +2,13 @@ Sastre 1.27 [XXX, 2026]
 ================================
 
 #### Improvements:
-- Added support for (device) tags.
+- Added support for device tags.
 - Added new show realtime bfd summary command
 - Added support for API key based SD-WAN manager authentication.
 
 #### Fixes:
 - Embedded-security profiles did not support all parcels available in 20.15. This is now fixed.
+- Fix for config-group deployment error with optional variables. 
 
 #### Changes:
 - Updated minimum Python version to 3.10 (from 3.9). As Python 3.9 is now EOL.